+++ CHANGELOG.md
@@ -260,3 +260,4 @@
 * Add in ``softban`` command.
 * Add in ``discordjs`` docs command.
 * Typescript now transpiles to ESNext, letting us use all the new keywords!
+* Remove ``https-proxy-agent`` and ``node-abort-controller`` in favor of native functions.